chore: remove deprecated classes
