--- src/abi.ts
@@ -1,9 +1,0 @@
-import AzraelAbi from './abi/azrael.abi';
-import SylvesterAbi from './abi/sylvester.abi';
-import ResolverAbi from './abi/resolver.abi';
-import WhoopiAbi from './abi/whoopi.abi';
-
-export const Azrael = AzraelAbi.abi;
-export const Sylvester = SylvesterAbi.abi;
-export const Resolver = ResolverAbi.abi;
-export const Whoopi = WhoopiAbi.abi;
