--- src/abi/azrael.abi.ts
@@ -1,9 +1,0 @@
-import {getDeploymentAbi,} from '../consts';
-import {AzraelVersion, RenftContractType} from "../types";
-
-export default {
-  abi: getDeploymentAbi({
-    contractType: RenftContractType.AZRAEL,
-    version: AzraelVersion.V0,
-  }),
-};
