--- src/abi/resolver.abi.ts
@@ -1,9 +1,0 @@
-import {getDeploymentAbi} from '../consts';
-import {RenftContractType, ResolverVersion} from '../types';
-
-export default {
-  abi: getDeploymentAbi({
-    contractType: RenftContractType.RESOLVER,
-    version: ResolverVersion.V0,
-  }),
-};
