--- src/abi/sylvester.abi.ts
@@ -1,9 +1,0 @@
-import {getDeploymentAbi} from '../consts';
-import {RenftContractType, SylvesterVersion} from '../types';
-
-export default {
-  abi: getDeploymentAbi({
-    contractType: RenftContractType.SYLVESTER,
-    version: SylvesterVersion.V0,
-  }),
-};
