+++ src/consts.ts
@@ -1,9 +1,9 @@
-import {Signer} from '@ethersproject/abstract-signer';
-import {Contract, ContractInterface} from '@ethersproject/contracts';
-import isEqual from 'react-fast-compare';
+import azrael_v0 from './abi/azrael.v0.abi.json';
+import sylvester_v0 from './abi/sylvester.v0.abi.json';
+import whoopi_v0 from './abi/whoopi.v0.abi.json';
+import resolver_v0 from './abi/resolver.v0.abi.json';
 
 import {
-  AbstractRenftContractDeployment,
   AzraelAbiVersions,
   AzraelVersion,
   ContractAbiVersions,
@@ -11,11 +11,8 @@ import {
   EVMNetworkLike,
   PaymentToken,
   PaymentTokenDetails,
-  RenftAzraelDeployment,
-  RenftContractDeployment,
-  RenftContractDeployments,
   RenftContracts,
-  RenftContractType, RenftResolverDeployment, RenftSylvesterDeployment, RenftWhoopiDeployment,
+  RenftContractType,
   ResolverAbiVersions,
   ResolverVersion,
   SylvesterAbiVersions,
@@ -23,12 +20,6 @@ import {
   WhoopiAbiVersions,
   WhoopiVersion,
 } from './types';
-
-import azrael_v0 from './abi/azrael.v0.abi.json';
-import sylvester_v0 from './abi/sylvester.v0.abi.json';
-import whoopi_v0 from './abi/whoopi.v0.abi.json';
-import resolver_v0 from './abi/resolver.v0.abi.json';
-import {createInterfaceVersions, CreateVersionedContractInterfaceResult} from "./contracts2";
 
 export const NETWORK_ETHEREUM_MAINNET: EVMNetworkLike<
   EVMNetworkType.ETHEREUM_MAINNET
@@ -58,81 +49,6 @@ export const NETWORK_AVALANCHE_MAINNET: EVMNetworkLike<
   chainId: 43_114,
 };
 
-export const DEPLOYMENT_AZRAEL_ETHEREUM_MAINNET_V0: RenftAzraelDeployment = {
-  contractAddress: '0x94d8f036a0fbc216bb532d33bdf6564157af0cd7',
-  network: NETWORK_ETHEREUM_MAINNET,
-  contractType: RenftContractType.AZRAEL,
-  version: AzraelVersion.V0,
-};
-
-export const DEPLOYMENT_SYLVESTER_ETHEREUM_MAINNET_V0: RenftSylvesterDeployment = {
-  contractAddress: '0xa8D3F65b6E2922fED1430b77aC2b557e1fa8DA4a',
-  network: NETWORK_ETHEREUM_MAINNET,
-  contractType: RenftContractType.SYLVESTER,
-  version: SylvesterVersion.V0,
-};
-
-export const DEPLOYMENT_SYLVESTER_POLYGON_MAINNET_V0: RenftSylvesterDeployment = {
-  contractAddress: '0xfA06cFE34C85Ec6b6D29A6a99806cC68BA0018Fe',
-  network: NETWORK_POLYGON_MAINNET,
-  contractType: RenftContractType.SYLVESTER,
-  version: SylvesterVersion.V0,
-};
-
-export const DEPLOYMENT_WHOOPI_AVALANCHE_FUJI_TESTNET_V0: RenftWhoopiDeployment = {
-  contractAddress: '0x42816FA3cB0aDc3fcAdED3109323c0Bc19215084',
-  network: NETWORK_AVALANCHE_FUJI_TESTNET,
-  contractType: RenftContractType.WHOOPI,
-  version: WhoopiVersion.V0,
-};
-
-export const DEPLOYMENT_WHOOPI_AVALANCHE_MAINNET_V0: RenftWhoopiDeployment = {
-  contractAddress: '0x6Ee495ecEd3A0255057667FF2685e53f54A19A65',
-  network: NETWORK_AVALANCHE_MAINNET,
-  contractType: RenftContractType.WHOOPI,
-  version: WhoopiVersion.V0,
-};
-
-export const DEPLOYMENT_RESOLVER_ETHEREUM_MAINNET_V0: RenftResolverDeployment = {
-  contractAddress: '0x945e589a4715d1915e6fe14f08e4887bc4019341',
-  network: NETWORK_ETHEREUM_MAINNET,
-  contractType: RenftContractType.RESOLVER,
-  version: ResolverVersion.V0,
-};
-
-export const DEPLOYMENT_RESOLVER_POLYGON_MAINNET_V0: RenftResolverDeployment = {
-  contractAddress: '0x6884d88Ce56C5C93F46eE23684eBA8628c90B518',
-  network: NETWORK_POLYGON_MAINNET,
-  contractType: RenftContractType.RESOLVER,
-  version: ResolverVersion.V0,
-};
-
-export const DEPLOYMENT_RESOLVER_AVALANCHE_FUJI_TESTNET_V0: RenftResolverDeployment = {
-  contractAddress: '0x23F7F8B03BAF01D5124255fE240E81BbBd3AEc0D',
-  network: NETWORK_AVALANCHE_FUJI_TESTNET,
-  contractType: RenftContractType.RESOLVER,
-  version: ResolverVersion.V0,
-};
-
-export const DEPLOYMENT_RESOLVER_AVALANCHE_MAINNET_V0: RenftResolverDeployment = {
-  contractAddress: '0xEBFd584AAC21dfEFF02c3d4f308B0962610a028A',
-  network: NETWORK_AVALANCHE_MAINNET,
-  contractType: RenftContractType.RESOLVER,
-  version: ResolverVersion.V0,
-};
-
-export const RENFT_CONTRACT_DEPLOYMENTS: RenftContractDeployments = [
-  DEPLOYMENT_AZRAEL_ETHEREUM_MAINNET_V0,
-  DEPLOYMENT_SYLVESTER_ETHEREUM_MAINNET_V0,
-  DEPLOYMENT_SYLVESTER_POLYGON_MAINNET_V0,
-  DEPLOYMENT_WHOOPI_AVALANCHE_FUJI_TESTNET_V0,
-  DEPLOYMENT_WHOOPI_AVALANCHE_MAINNET_V0,
-  DEPLOYMENT_RESOLVER_ETHEREUM_MAINNET_V0,
-  DEPLOYMENT_RESOLVER_POLYGON_MAINNET_V0,
-  DEPLOYMENT_RESOLVER_AVALANCHE_FUJI_TESTNET_V0,
-  DEPLOYMENT_RESOLVER_AVALANCHE_MAINNET_V0,
-];
-
 export const AZRAEL_ABI_VERSIONS: AzraelAbiVersions = {
   [AzraelVersion.V0]: azrael_v0,
 };
@@ -156,173 +72,10 @@ export const CONTRACT_ABI_VERSIONS: ContractAbiVersions = {
   [RenftContractType.RESOLVER]: RESOLVER_ABI_VERSIONS,
 };
 
-// Attempts to find matching deployments for a given set of deployment properties.
-// For example, it allows callers to search for all deployments for a given network,
-// or of a specific contractType and version.
-export function findDeployments<T extends RenftContractDeployment>(search: Partial<T>) {
-  return RENFT_CONTRACT_DEPLOYMENTS
-    .filter(
-      (maybeMatchingDeployment: RenftContractDeployment): maybeMatchingDeployment is T  => {
-        const definedKeys = Object.keys(search);
-        const filterObject = Object
-          .fromEntries(
-            Object.entries(maybeMatchingDeployment)
-                  .filter(([k]) => definedKeys.includes(k)),
-          );
-        return isEqual(filterObject, search);
-      },
-    );
-}
-
-// Find a single contract address for a given deployment. Will throw if none-or-many
-// matching deployments are found.
-export function getContractAddressForDeployment<T extends RenftContractDeployment>(
-  search: Omit<Partial<T>, 'contractAddress'>,
-): string {
-  const matchingDeployments = findDeployments<T>(search as Partial<T>);
-
-  if (!matchingDeployments.length)
-    throw new Error(
-      `[getContractAddressForDeployment]: Failed to find a matching deployment for search: ${
-        JSON.stringify(search)
-      }`
-    );
-
-  if (matchingDeployments.length > 1)
-    throw new Error(
-      `[getContractAddressForDeployment]: Found multiple possible deployments for search: ${
-        JSON.stringify(search)
-      }`
-    );
-
-  const [matchingDeployment] = matchingDeployments;
-
-  const {contractAddress} = matchingDeployment!;
-
-  return contractAddress;
-}
-
-export function getDeploymentAbi<T extends RenftContractType>({
-  contractType,
-  version,
-}: {
-  readonly contractType: T;
-  readonly version: keyof CreateVersionedContractInterfaceResult[T];
-}): ContractInterface {
-  const contractAbiVersions = CONTRACT_ABI_VERSIONS[contractType];
-
-  // @ts-expect-error versioning isn't deterministic at this point because we haven't strictly typed AbiVersions with InterfaceVersions.
-  const maybeContractAbi = contractAbiVersions[version];
-
-  if (!maybeContractAbi)
-    throw new Error(
-      `[getDeploymentAbi]: Unable to find abi for combination "${
-        String(contractType)
-      }", "${
-        String(version)
-      }".`,
-    );
-
-  return maybeContractAbi;
-}
-
-export function getContractForDeployment<T extends RenftContractType>({
-  contractAddress,
-  contractType,
-  version,
-  signer,
-}: {
-  readonly contractAddress: string;
-  readonly contractType: T;
-  readonly version: keyof CreateVersionedContractInterfaceResult[T];
-  readonly signer: Signer;
-}): Contract {
-
-  const abi = getDeploymentAbi({contractType, version});
-
-  return new Contract(contractAddress, abi, signer);
-}
-
-export function getVersionedContractInterfaceForDeployment<
-  ContractType extends keyof CreateVersionedContractInterfaceResult,
-  Version extends keyof CreateVersionedContractInterfaceResult[ContractType],
->({
-  deployment,
-  signer,
-}: {
-  readonly deployment: AbstractRenftContractDeployment<ContractType, Version>;
-  readonly signer: Signer;
-}) {
-  const {
-    contractAddress,
-    contractType,
-    version,
-  } = deployment;
-
-  const contract = getContractForDeployment({
-    contractAddress,
-    contractType,
-    version,
-    signer,
-  });
-
-  const {
-    [contractType]: contractFunctions
-  } = createInterfaceVersions(contract);
-
-  return contractFunctions[version];
-}
-
 // TODO: enforce this relationship with ContractTypes
 // * Note, this price does not apply to Whoopi
 export const MAX_PRICE = 9999.9999;
 export const NUM_BITS_IN_BYTE = 8;
-
-export const ResolverAddress = getContractAddressForDeployment({
-  network: NETWORK_ETHEREUM_MAINNET,
-  contractType: RenftContractType.RESOLVER,
-});
-
-export const ResolverPolygonAddress = getContractAddressForDeployment({
-  network: NETWORK_POLYGON_MAINNET,
-  contractType: RenftContractType.RESOLVER,
-});
-
-export const ResolverFujiAddress = getContractAddressForDeployment({
-  network: NETWORK_AVALANCHE_FUJI_TESTNET,
-  contractType: RenftContractType.RESOLVER,
-});
-
-export const ResolverAvalancheAddress = getContractAddressForDeployment({
-  network: NETWORK_AVALANCHE_MAINNET,
-  contractType: RenftContractType.RESOLVER,
-});
-
-// TODO: DEPRECATE THESE WITH WARNING
-export const AzraelAddress = getContractAddressForDeployment({
-  network: NETWORK_ETHEREUM_MAINNET,
-  contractType: RenftContractType.AZRAEL,
-});
-
-export const SylvesterAddress = getContractAddressForDeployment({
-  network: NETWORK_ETHEREUM_MAINNET,
-  contractType: RenftContractType.SYLVESTER,
-});
-
-export const SylvesterPolygonAddress = getContractAddressForDeployment({
-  network: NETWORK_POLYGON_MAINNET,
-  contractType: RenftContractType.SYLVESTER,
-});
-
-export const WhoopiFujiAddress = getContractAddressForDeployment({
-  network: NETWORK_AVALANCHE_FUJI_TESTNET,
-  contractType: RenftContractType.WHOOPI,
-});
-
-export const WhoopiAvalancheAddress = getContractAddressForDeployment({
-  network: NETWORK_AVALANCHE_MAINNET,
-  contractType: RenftContractType.WHOOPI,
-});
 
 // Resolver related
 const SENTINEL: PaymentTokenDetails = {
