+++ src/contracts/azrael/index.ts
@@ -1,0 +1,2 @@
+export * from './types';
+export * from './utils';
