+++ test/consts.test.ts
@@ -1,3 +1,4 @@
+import {Signer} from '@ethersproject/abstract-signer';
 import { expect } from 'chai';
 
 import {
@@ -10,17 +11,14 @@ import {
   WHOOPI_AVALANCHE_ADDRESS,
   WHOOPI_FUJI_ADDRESS,
   RESOLVER_FUJI_ADDRESS,
-} from '../src';
-import {
   DEPLOYMENT_AZRAEL_ETHEREUM_MAINNET_V0,
   DEPLOYMENT_SYLVESTER_ETHEREUM_MAINNET_V0,
   DEPLOYMENT_WHOOPI_AVALANCHE_MAINNET_V0,
-  getVersionedContractInterfaceForDeployment
-} from '../src/consts';
-import {AzraelV0FunctionInterface} from "../src/contracts2/azrael/types";
-import {Signer} from "@ethersproject/abstract-signer";
-import {SylvesterV0FunctionInterface} from "../src/contracts2/sylvester/types";
-import {WhoopiV0FunctionInterface} from "../src/contracts2/whoopi/types";
+  getVersionedContractInterfaceForDeployment,
+  AzraelV0FunctionInterface,
+  SylvesterV0FunctionInterface,
+  WhoopiV0FunctionInterface,
+} from '../src';
 
 describe('deprecated contract addresses', () => {
   it('should not invalidate current consumers', () => {
